Public API

LDR Factorization

StableLinearAlgebra.LDR — Type

LDR{T<:Number, E<:Real} <: Factorization{T}

Represents the matrix factorization $A P = L D R$ for a square matrix $A,$ which may equivalently be written as $A = (L D R) P^{-1} = (L D R) P^T$.

In the above $L$ is a unitary matrix, $D$ is a diagonal matrix of strictly positive real numbers, and $R$ is an upper triangular matrix. Lastly, $P$ is a permutation matrix, for which $P^{-1}=P^T$.

This factorization is based on a column-pivoted QR decomposition $A P = Q R,$ such that

\[\begin{align*} L &= Q \\ D &= \vert \textrm{diag}(R) \vert \\ R &= \vert \textrm{diag}(R) \vert^{-1} R \\ P &= P. \end{align*}\]

Fields

L::Matrix{T}: The unitary matrix $L.$

d::Vector{E}: A vector representing the diagonal matrix $D.$

R::Matrix{T}: The upper triangular matrix $R.$

pᵀ::Vector{Int}: A permutation vector representing the permuation matrix $P^T.$

StableLinearAlgebra.ldr — Function

ldr(A::AbstractMatrix)

Calculate and return the LDR decomposition for the matrix A.

ldr(F::LDR)

Return a new LDR factorization that is a copy of F.

StableLinearAlgebra.ldr! — Function

ldr!(F::LDR, A::AbstractMatrix)

Calculate the LDR decomposition F for the matrix A.

ldr!(F::LDR)

Re-calculate the LDR factorization F in-place based on the current contents of the matrix F.L.

ldr!(F::LDR, I::UniformScaling)

Update the LDR factorization F to reflect the identity matrix.

StableLinearAlgebra.ldrs — Function

ldrs(A::AbstractMatrix{T}, N::Int) where {T}

Return a vector of N LDR factorizations, where each one represent the matrix A.

ldrs(A::AbstractArray{T,3}) where {T}

Return a vector of size(A, 3) LDR factorizations, where there is an LDR factorization for each matrix A[:,:,i].

StableLinearAlgebra.ldrs! — Function

ldrs!(Fs::Vector{LDR{T}}, A::AbstractArray{T,3}) where {T}

Update the vector Fs of LDR factorization based on the sequence of matrices contained in A.

StableLinearAlgebra.LDRWorkspace — Type

LDRWorkspace{T<:Number}

A workspace to avoid temporary memory allocations when performing computations with an LDR factorization.

StableLinearAlgebra.ldr_workspace — Function

ldr_workspace(A::AbstractMatrix)
ldr_workspace(F::LDR)
ldr_workspace(Fs::Vector{LDR})

Return a workspace for an LDR factorization that can be used to avoid dynamic memory allocations.

Overloaded Functions

Base.size — Function

size(F::LDR)
size(F::LDR, dims)

Return the size of the LDR decomposition F.

Base.copyto! — Function

copyto!(A::AbstractMatrix{T}, F::LDR{T}, ws::LDRWorkspace{T}) where {T}
copyto!(A::AbstractMatrix{T}, F::LDR{T};
        M{T}::AbstractMatrix=similar(A)) where {T}

Copy the matrix represented by the LDR decomposition F into the matrix A.

copyto!(F::LDR{T}, A::AbstractMatrix{T}) where {T}

Copy the matrix A to the LDR factorization F, calculating the LDR factorization to represent A.

copyto!(F′::LDR{T}, F::LDR{T}) where {T}

Copy F LDR factorization to F′.

LinearAlgebra.mul! — Function

mul!(A::AbstractMatrix{T}, F::LDR{T}, B::AbstractMatrix{T}, ws::LDRWorkspace{T}) where {T}
mul!(A::AbstractMatrix{T}, F::LDR{T}, B::AbstractMatrix{T};
     M::AbstractMatrix{T} = similar(A)) where {T}

Calculate the matrix product $A = M B,$ where the matrix $M$ is represented by the LDR factorization F.

mul!(F′::LDR{T}, B::AbstractMatrix{T}, F::LDR{T}, ws::LDRWorkspace{T}) where {T}
mul!(F′::LDR{T}, B::AbstractMatrix{T}, F::LDR{T};
     M::AbstractMatrix{T}=similar(F.L)) where {T}

Calculate the numerically stable product $C = B A,$ where the matrices $C$ and $A$ are represented by the LDR decompositions F′ and F respectively.

Algorithm

Calculate the matrix product using the procedure

\[\begin{align*} C = & BA\\ = & \overset{L_{1}D_{1}R_{1}P_{1}^{T}}{\overbrace{B[L_{0}D_{0}}}R_{0}P_{0}^{T}]\\ = & [\overset{L_{2}}{\overbrace{L_{1}^{\phantom{\dagger}}}}\,\overset{D_{2}}{\overbrace{D_{1}^{\phantom{\dagger}}}}\,\overset{R_{2}}{\overbrace{R_{1}P_{1}^{T}]R_{0}}}\,\overset{P_{2}^{T}}{\overbrace{P_{0}^{T}}}\\ = & L_{2}D_{2}R_{2}P_{2}^{T}. \end{align*}\]

mul!(F′::LDR{T}, F::LDR{T}, B::AbstractMatrix{T}, ws::LDRWorkspace{T}) where {T}
mul!(F′::LDR{T}, F::LDR{T}, B::AbstractMatrix{T};
     M::AbstractMatrix{T}=similar(B)) where {T}

Calculate the numerically stable product $C = A B,$ where the matrices $C$ and $A$ are represented by the LDR decompositions F′ and F respectively.

Algorithm

Calculate the matrix product using the procedure

\[\begin{align*} C = & AB\\ = & [L_{0}\overset{L_{1}D_{1}R_{1}P_{1}^{T}}{\overbrace{D_{0}R_{0}P_{0}^{T}]B}}\\ = & \overset{L_{2}}{\overbrace{L_{0}[L_{1}^{\phantom{T}}}}\,\overset{D_{2}}{\overbrace{D_{1}^{\phantom{T}}}}\,\overset{R_{2}}{\overbrace{R_{1}^{\phantom{T}}}}\,\overset{P_{2}^{T}}{\overbrace{P_{1}^{T}}}]\\ = & L_{2}D_{2}R_{2}P_{2}^{T}. \end{align*}\]

mul!(F₃::LDR{T}, F₂::LDR{T}, F₁::LDR{T}, ws::LDRWorkspace{T}) where {T}
mul!(F₃::LDR{T}, F₂::LDR{T}, F₁::LDR{T};
     M::AbstractMatrix{T}) where {T}

Calculate the numerically stable product $A = B C,$ where each matrix is represented by the LDR decompositions F₃, F₂ and F₁ respectively.

Algorithm

Calculate the numerically stable matrix product $A = B C$ using the procedure

\[\begin{align*} A = & BC\\ = & [L_{b}D_{b}\overset{M}{\overbrace{R_{b}P_{b}^{T}][L_{c}}}D_{c}R_{c}P_{c}^{T}]\\ = & L_{b}D_{b}\overset{M'}{\overbrace{MD_{c}}}R_{c}P_{c}^{T}\\ = & L_{b}\overset{M''}{\overbrace{D_{b}M'}}R_{c}P_{c}^{T}\\ = & \overset{L_{0}D_{0}R_{0}P_{0}^{T}}{L_{b}\overbrace{M''}R_{c}}P_{c}^{T}\\ = & \overset{L_{a}}{\overbrace{L_{b}L_{0}^{\phantom{T}}}}\,\overset{D_{a}}{\overbrace{D_{0}^{\phantom{T}}}}\,\overset{R_{a}}{\overbrace{R_{0}P_{0}^{T}R_{c}}}\,\overset{P_{a}^{T}}{\overbrace{P_{c}^{T}}}\\ = & L_{a}D_{a}R_{a}P_{a}^{T}. \end{align*}\]

LinearAlgebra.lmul! — Function

lmul!(B::AbstractMatrix{T}, F::LDR{T}, ws::LDRWorkspace{T}) where {T}
lmul!(B::AbstractMatrix{T}, F::LDR{T};
      M::AbstractMatrix{T}=similar(B),
      M′::AbstractMatrix{T}=similar(B),
      p::AbstractMatrix{Int}=similar(F.pᵀ)) where {T}

Calculate the numerically stable product $A = B A,$ where the matrix $A$ is represented by the LDR factorization F, updating F in-place to represent the product $B A.$

Algorithm

Given a matrix $A$ represented by an LDR factorization, update the LDR factorization to reflect the matrix product $A=BA$ using the procedure

\[\begin{align*} A = & BA\\ = & \overset{L_{1}D_{1}R_{1}P_{1}^{T}}{\overbrace{B[L_{0}D_{0}}}R_{0}P_{0}^{T}]\\ = & [\overset{L_{2}}{\overbrace{L_{1}^{\phantom{\dagger}}}}\,\overset{D_{2}}{\overbrace{D_{1}^{\phantom{\dagger}}}}\,\overset{R_{2}}{\overbrace{R_{1}P_{1}^{T}]R_{0}}}\,\overset{P_{2}^{T}}{\overbrace{P_{0}^{T}}}\\ = & L_{2}D_{2}R_{2}P_{2}^{T}. \end{align*}\]

lmul!(F₂::LDR{T}, F₁::LDR{T}, ws::LDRWorkspace{T}) where {T}
lmul!(F₂::LDR{T}, F₁::LDR{T};
      M::AbstractMatrix{T}=similar(F.L),
      M′::AbstractMatrix{T}=similar(F.L),
      p::AbstractVector{Int}=similar(F₁.pᵀ)) where {T}

Calculate the matrix product $C = B C,$ represented by the LDR factorization F₂ and F₁ respectively, where F₁ is updated in-place.

Algorithm

Calculate the numerically stable matrix product $C = B C$ using the procedure

\[\begin{align*} C = & B C\\ = & [L_{b}D_{b}\overset{M}{\overbrace{R_{b}P_{b}^{T}][L_{c}}}D_{c}R_{c}P_{c}^{T}]\\ = & L_{b}D_{b}\overset{M'}{\overbrace{MD_{c}}}R_{c}P_{c}^{T}\\ = & L_{b}\overset{M''}{\overbrace{D_{b}M'}}R_{c}P_{c}^{T}\\ = & \overset{L_{0}D_{0}R_{0}P_{0}^{T}}{L_{b}\overbrace{M''}R_{c}}P_{c}^{T}\\ = & \overset{L_{c}}{\overbrace{L_{c}L_{0}^{\phantom{T}}}}\,\overset{D_{c}}{\overbrace{D_{0}^{\phantom{T}}}}\,\overset{R_{c}}{\overbrace{R_{0}P_{0}^{T}R_{c}}}\,\overset{P_{c}^{T}}{\overbrace{P_{c}^{T}}}\\ = & L_{c}D_{c}R_{c}P_{c}^{T}. \end{align*}\]

LinearAlgebra.rmul! — Function

rmul!(F::LDR{T}, B::AbstractMatrix{T}, ws::LDRWorkspace{T}) where {T}
rmul!(F::LDR{T}, B::AbstractMatrix{T};
      M::AbstractMatrix{T}=similar(B),
      M′::AbstractMatrix{T}=similar(B)) where {T}

Calculate the numerically stable product $A = A B,$ where the matrix $A$ is represented by the LDR factorization F, updating F in-place to represent the product $A B.$

Algorithm

Given a matrix $A$ represented by an LDR factorization, update the LDR factorization to reflect the matrix product $A=AB$ using the procedure

\[\begin{align*} A = & AB\\ = & [L_{0}\overset{L_{1}D_{1}R_{1}P_{1}^{T}}{\overbrace{D_{0}R_{0}P_{0}^{T}]B}}\\ = & \overset{L_{2}}{\overbrace{L_{0}[L_{1}^{\phantom{T}}}}\,\overset{D_{2}}{\overbrace{D_{1}^{\phantom{T}}}}\,\overset{R_{2}}{\overbrace{R_{1}^{\phantom{T}}}}\,\overset{P_{2}^{T}}{\overbrace{P_{1}^{T}}}]\\ = & L_{2}D_{2}R_{2}P_{2}^{T}. \end{align*}\]

rmul!(F₂::LDR{T}, F₁::LDR{T}, ws::LDRWorkspace{T}) where {T}
rmul!(F₂::LDR{T}, F₁::LDR{T};
      M::AbstractMatrix{T}=similar(F.L),
      M′::AbstractMatrix{T}=similar(F.L)) where {T}

Calculate the matrix product $B = B C,$ represented by the LDR factorization F₂ and F₁ respectively, where F₂ is updated in-place.

Algorithm

Calculate the numerically stable matrix product $B = B C$ using the procedure

\[\begin{align*} B = & B C\\ = & [L_{b}D_{b}\overset{M}{\overbrace{R_{b}P_{b}^{T}][L_{c}}}D_{c}R_{c}P_{c}^{T}]\\ = & L_{b}D_{b}\overset{M'}{\overbrace{MD_{c}}}R_{c}P_{c}^{T}\\ = & L_{b}\overset{M''}{\overbrace{D_{b}M'}}R_{c}P_{c}^{T}\\ = & \overset{L_{0}D_{0}R_{0}P_{0}^{T}}{L_{b}\overbrace{M''}R_{c}}P_{c}^{T}\\ = & \overset{L_{b}}{\overbrace{L_{b}L_{0}^{\phantom{T}}}}\,\overset{D_{b}}{\overbrace{D_{0}^{\phantom{T}}}}\,\overset{R_{b}}{\overbrace{R_{0}P_{0}^{T}R_{c}}}\,\overset{P_{b}^{T}}{\overbrace{P_{c}^{T}}}\\ = & L_{b}D_{b}R_{b}P_{b}^{T}. \end{align*}\]

LinearAlgebra.det — Function

det(F::LDR{T}, ws::LDRWorkspace{T}) where {T}
det(F::LDR{T}) where {T<:Real}
det(F::LDR{T}; M::AbstractMatrix{T}=similar(F.L)) where {T<:Complex}

Return the determinant of the LDR factorization F.

Exported Functions

StableLinearAlgebra.inv! — Function

inv!(A⁻¹::AbstractMatrix{T}, F::LDR{T}, ws::LDRWorkspace{T}) where {T}
inv!(A⁻¹::AbstractMatrix{T}, F::LDR{T};
     M::AbstractMatrix{T}=similar(F.L),
     p::AbstractVector{Int}=similar(F.pᵀ)) where {T}

Calculate the inverse of a matrix $A$ represented of the LDR decomposition F, writing the inverse matrix A⁻¹.

inv!(F::LDR{T}, ws::LDRWorkspace{T}) where {T}
inv!(F::LDR{T};
     M::AbstractMatrix{T},
     p::AbstractVector{Int}) where {T}

Invert the LDR decomposition F in-place.

StableLinearAlgebra.inv_IpA! — Function

inv_IpA!(G::AbstractMatrix{T}, F::LDR{T}, ws::LDRWorkspace{T};
         F′::LDR{T}=ldr(F)) where {T}
inv_IpA!(G::AbstractMatrix{T}, F::LDR{T};
         F′::LDR{T}=ldr(F),
         d_min::AbstractVector{T}=similar(F.d),
         d_max::AbstractVector{T}=similar(F.d),
         M::AbstractMatrix{T}=similar(F.L),
         p::AbstractVector{Int}=similar(F.pᵀ)) where {T}

Given a matrix $A$ represented by the LDR factorization F, calculate the numerically stabalized inverse

\[G = (I + A)^{-1},\]

storing the result in the matrix G.

Algorithm

Given an LDR factorization of $A$, calculate $G = (I + A)^{-1}$ using the procedure

\[\begin{align*} G = & \left(I+A\right)^{-1}\\ = & (I+\overset{D_{a,\min}D_{a,\max}}{L_{a}\overbrace{D_{a}}R_{a}}P_{a}^{T})^{-1}\\ = & \left(I+L_{a}D_{a,\min}D_{a,\max}R_{a}P_{a}^{T}\right)^{-1}\\ = & \left(\left[P_{a}R_{a}^{-1}D_{a,\max}^{-1}+L_{a}D_{a,\min}\right]D_{a,\max}R_{a}P_{a}^{T}\right)^{-1}\\ = & P_{a}R_{a}^{-1}D_{a,\max}^{-1}(\overset{L_{0}D_{0}R_{0}P_{0}^{T}}{\overbrace{P_{a}R_{a}^{-1}D_{a,\max}^{-1}+L_{a}D_{a,\min}}})^{-1}\\ = & P_{a}R_{a}^{-1}D_{a,\max}^{-1}(L_{0}D_{0}R_{0}P_{0}^{T})^{-1}\\ = & P_{a}R_{a}^{-1}\overset{L_{1}D_{1}R_{1}P_{1}^{T}}{\overbrace{D_{a,\max}^{-1}P_{0}R_{0}^{-1}D_{0}^{-1}L_{0}^{\dagger}}}\\ = & P_{a}R_{a}^{-1}L_{1}D_{1}R_{1}P_{1}^{T}, \end{align*}\]

where $D_{\min} = \min(D, 1)$ and $D_{\max} = \max(D, 1).$

StableLinearAlgebra.inv_UpV! — Function

inv_UpV!(G::AbstractMatrix{T}, Fᵤ::LDR{T}, Fᵥ::LDR{T}, ws::LDRWorkspace{T};
         F::LDR{T}=ldr(Fᵥ)) where {T}
inv_UpV!(G::AbstractMatrix{T}, Fᵤ::LDR{T}, Fᵥ::LDR{T};
         F::LDR{T}=ldr(Fᵤ),
         dᵤ_min::AbstractVector{T}=similar(Fᵤ.d),
         dᵤ_max::AbstractVector{T}=similar(Fᵤ.d),
         dᵥ_min::AbstractVector{T}=similar(Fᵥ.d),
         dᵥ_max::AbstractVector{T}=similar(Fᵥ.d),
         M::AbstractMatrix{T}=similar(Fᵥ.L),
         M′::AbstractMatrix{T}=similar(Fᵥ.L),
         M″::AbstractMatrix{T}=similar(Fᵥ.L),
         p::AbstractVector{Int}=similar(Fᵥ.pᵀ),
         p′::AbstractVector{Int}=similar(Fᵥ.pᵀ)) where {T}

Calculate the numerically stable inverse $G = (U + V)^{-1},$ where the matrices $U$ and $V$ are represented by the LDR factorizations Fᵤ and Fᵥ respectively.

Algorithm

Letting $U = [L_u D_u R_u] P_u^T$ and $V = [L_v D_v R_v] P_v^T,$ the inverse matrix $G = (U + V)^{-1}$ is calculated using the procedure

\[\begin{align*} G = & \left(U+V\right)^{-1}\\ = & \overset{D_{u,\max}D_{u,\min}}{(L_{u}\overbrace{D_{u}}R_{u}}P_{u}^{T}+\overset{D_{v,\min}D_{v,\max}}{L_{v}\overbrace{D_{v}}R_{v}}P_{v}^{T})^{-1}\\ = & \left(L_{u}D_{u,\max}D_{u,\min}R_{u}P_{u}^{T}+L_{v}D_{v,\min}D_{v,\max}R_{v}P_{v}^{T}\right)^{-1}\\ = & \left(L_{u}D_{u,\max}\left[D_{u,\min}R_{u}P_{u}^{T}P_{v}R_{v}^{-1}D_{v,\max}^{-1}+D_{u,\max}^{-1}L_{u}^{\dagger}L_{v}D_{v,\min}\right]D_{v,\max}R_{v}P_{v}^{T}\right)^{-1}\\ = & P_{v}R_{v}^{-1}D_{v,\max}^{-1}(\overset{L_{0}D_{0}R_{0}P_{0}^{T}}{\overbrace{D_{u,\min}R_{u}P_{u}^{T}P_{v}R_{v}^{-1}D_{v,\max}^{-1}+D_{u,\max}^{-1}L_{u}^{\dagger}L_{v}D_{v,\min}}})^{-1}D_{u,\max}^{-1}L_{u}^{\dagger}\\ = & P_{v}R_{v}^{-1}D_{v,\max}^{-1}\left(L_{0}D_{0}R_{0}P_{0}^{T}\right)^{-1}D_{u,\max}^{-1}L_{u}^{\dagger}\\ = & P_{v}R_{v}^{-1}\overset{L_{1}D_{1}R_{1}P_{1}^{T}}{\overbrace{D_{v,\max}^{-1}P_{0}R_{0}^{-1}D_{0}^{-1}L_{0}^{\dagger}D_{u,\max}^{-1}}}L_{u}^{\dagger}\\ = & P_{v}R_{v}^{-1}L_{1}D_{1}R_{1}P_{1}^{T}L_{u}^{\dagger}, \end{align*}\]

where $D_\textrm{min} = \min(D,1)$ and $D_\textrm{max} = \max(D,1).$

StableLinearAlgebra.inv_invUpV! — Function

inv_invUpV!(G::AbstractMatrix{T}, Fᵤ::LDR{T}, Fᵥ::LDR{T}, ws::LDRWorkspace{T};
            F::LDR{T}=ldr(F)) where {T}
inv_invUpV!(G::AbstractMatrix{T}, Fᵤ::LDR{T}, Fᵥ::LDR{T};
            F::LDR{T}=ldr(Fᵤ),
            dᵤ_min::AbstractVector{T}=similar(Fᵤ.d),
            dᵤ_max::AbstractVector{T}=similar(Fᵤ.d),
            dᵥ_min::AbstractVector{T}=similar(Fᵥ.d),
            dᵥ_max::AbstractVector{T}=similar(Fᵥ.d),
            M::AbstractMatrix{T}=similar(Fᵥ.L),
            M′::AbstractMatrix{T}=similar(Fᵥ.L),
            p::AbstractVector{Int}=similar(Fᵥ.pᵀ),
            p′::AbstractVector{Int}=similar(Fᵥ.pᵀ)) where {T}

Calculate the numerically stable inverse $G = (U^{-1} + V)^{-1},$ where the matrices $U$ and $V$ are represented by the LDR factorizations Fᵤ and Fᵥ respectively.

Algorithm

Letting $U = [L_u D_u R_u] P_u^T$ and $V = [L_v D_v R_v] P_v^T,$ the inverse matrix $G = (U^{-1} + V)^{-1}$ is calculated using the procedure

\[\begin{align*} G = & \left(U^{-1}+V\right)^{-1}\\ = & ([\stackrel{D_{u,\max}D_{u,\min}}{L_{u}\overbrace{D_{u}}R_{u}}P_{u}^{T}]^{-1}+\overset{D_{v,\min}D_{v,\max}}{L_{v}\overbrace{D_{v}}R_{v}}P_{v}^{T})^{-1}\\ = & \left(\left[L_{u}D_{u,\max}D_{u,\min}R_{u}P_{u}^{T}\right]^{-1}+L_{v}D_{v,\min}D_{v,\max}R_{v}P_{v}^{T}\right)^{-1}\\ = & \left(P_{u}R_{u}^{-1}D_{u,\min}^{-1}D_{u,\max}^{-1}L_{u}^{\dagger}+L_{v}D_{v,\min}D_{v,\max}R_{v}P_{v}^{T}\right)^{-1}\\ = & \left(P_{u}R_{u}^{-1}D_{u,\min}^{-1}\left[D_{u,\max}^{-1}L_{u}^{\dagger}P_{v}R_{v}^{-1}D_{v,\max}^{-1}+D_{u,\min}R_{u}P_{u}^{T}L_{v}D_{v,\min}\right]D_{v,\max}R_{v}P_{v}^{T}\right)^{-1}\\ = & P_{v}R_{v}^{-1}D_{v,\max}^{-1}(\overset{L_{0}D_{0}R_{0}P_{0}^{T}}{\overbrace{D_{u,\max}^{-1}L_{u}^{\dagger}P_{v}R_{v}^{-1}D_{v,\max}^{-1}+D_{u,\min}R_{u}P_{u}^{T}L_{v}D_{v,\min}}})^{-1}D_{u,\min}R_{u}P_{u}^{T}\\ = & P_{v}R_{v}^{-1}D_{v,\max}^{-1}\left(L_{0}D_{0}R_{0}P_{0}^{T}\right)^{-1}D_{u,\min}R_{u}P_{u}^{T}\\ = & P_{v}R_{v}^{-1}\overset{L_{1}D_{1}R_{1}P_{1}^{T}}{\overbrace{D_{v,\max}^{-1}P_{0}R_{0}^{-1}D_{0}^{-1}L_{0}^{\dagger}D_{u,\min}}}R_{u}P_{u}^{T}\\ = & P_{v}R_{v}^{-1}L_{1}D_{1}R_{1}P_{1}^{T}R_{u}P_{u}^{T}, \end{align*}\]

where $D_\textrm{min} = \min(D,1)$ and $D_\textrm{max} = \max(D,1).$

StableLinearAlgebra.sign_det — Function

sign_det(F::LDR{T}, ws::LDRWorkspace{T}) where {T}
sign_det(F::LDR{T}) where {T<:Real}
sign_det(F::LDR{T}; M::AbstractMatrix{T}=similar(F.L)) where {T<:Complex}

Returns the sign/phase factor of the determinant for a matrix $A$ represented by the LDR factorization F, which is calculated as

\[\textrm{sgn}(\det A) = \det L \cdot \left( \prod_i R_{i,i} \right) \cdot \det P^T,\]

where $A = [L D R]P^T.$

StableLinearAlgebra.abs_det — Function

abs_det(F::LDR{T}; as_log::Bool=false) where {T}

Calculate the absolute value of determinant of the LDR factorization F. If as_log=true, then the log of the absolute value of the determinant is returned instead.

Algorithm

Given an LDR factorization $[L D R]P^T,$ calculate the absolute value of the determinant as

\[\exp\left\{ \sum_i \log(D[i]) \right\},\]

where $D$ is a diagonal matrix with strictly positive real matrix elements.

StableLinearAlgebra.abs_det_ratio — Function

abs_det_ratio(F₂::LDR{T}, F₁::LDR{T}, ws::LDRWorkspace{T};
              as_log::Bool=false) where {T}
abs_det_ratio(F₂::LDR{T}, F₁::LDR{T};
              as_log::Bool=false,
              p::AbstractVector{Int}=similar(F₁.pᵀ),
              p′::AbstractVector{Int}=similar(F₁.pᵀ)) where {T}

Given two matrices $A_2$ and $A_1$ represented by the LDR factorizations F₂ and F₁ respectively, calculate the absolute value of the determinant ratio $\vert\det(A_2/A_1)\vert$ in a numerically stable fashion. If as_log=true, then this function instead returns $\log \left( \vert\det(A_2/A_1)\vert \right).$

Algorithm

Let $A_1 = [L_1 D_1 R_1] P_1^T$ and $A_2 = [L_2 D_2 R_2] P_1^T$ be $N \times N$ square matrices each represented by their respective LDR factorizations. Let us define perumations $p_1^{(1)} \dots p_1^{(i)} \dots p_1^{(N)}$ and $p_2^{(1)} \dots p_2^{(i)} \dots p_2^{(N)}$ that sort the diagonal elements of $D_1$ and $D_2$ from smallest to largest. Then a numerically stable expression for evaulating the absolute value of the determinant ratio is

\[\vert \det(A_2/A_1) \vert = \exp\left\{ \sum_i \left( \log(D_2[p_2^{(i)}]) - \log(D_1[p_1^{(i)}]) \right) \right\},\]

keeping in mind that the diagonal elements of $D_1$ and $D_2$ are stictly positive real numbers.

Developer API

StableLinearAlgebra.mul_D! — Function

mul_D!(A, d, B)

Calculate the matrix product $A = D \cdot B,$ where $D$ is a diagonal matrix represented by the vector d.

mul_D!(A, B, d)

Calculate the matrix product $A = B \cdot D,$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.div_D! — Function

div_D!(A, d, B)

Calculate the matrix product $A = D^{-1} \cdot B,$ where $D$ is a diagonal matrix represented by the vector d.

div_D!(A, B, d)

Calculate the matrix product $A = B \cdot D^{-1},$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.lmul_D! — Function

lmul_D!(d, M)

In-place calculation of the matrix product $M = D \cdot M,$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.rmul_D! — Function

rmul_D!(M, d)

In-place calculation of the matrix product $M = M \cdot D,$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.ldiv_D! — Function

ldiv_D!(d, M)

In-place calculation of the matrix product $M = D^{-1} \cdot M,$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.rdiv_D! — Function

rdiv_D!(M, d)

In-place calculation of the matrix product $M = M \cdot D^{-1},$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.mul_P! — Function

mul_P!(A, p, B)

Evaluate the matrix product $A = P \cdot B,$ where $P$ is a permutation matrix represented by the vector of integers p.

mul_P!(A, B, p)

Evaluate the matrix product $A = B \cdot P,$ where $P$ is a permutation matrix represented by the vector of integers p.

StableLinearAlgebra.inv_P! — Function

inv_P!(p⁻¹, p)

Calculate the inverse/transpose $P^{-1}=P^T$ of a permuation $P$ represented by the vector p, writing the result to p⁻¹.

StableLinearAlgebra.sign_P — Function

perm_sign(p::AbstractVector{Int})

Calculate the sign/parity of the permutation, $\textrm{sgn}(p) = \pm 1.$