Developer API

StableLinearAlgebra.det_D — Function

det_D(d::AbstractVector{T}) where {T}

Given a diagonal matrix $D$ represented by the vector d, return $\textrm{sign}(\det D)$ and $\log(|\det A|).$

StableLinearAlgebra.mul_D! — Function

mul_D!(A, d, B)

Calculate the matrix product $A = D \cdot B,$ where $D$ is a diagonal matrix represented by the vector d.

mul_D!(A, B, d)

Calculate the matrix product $A = B \cdot D,$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.div_D! — Function

div_D!(A, d, B)

Calculate the matrix product $A = D^{-1} \cdot B,$ where $D$ is a diagonal matrix represented by the vector d.

div_D!(A, B, d)

Calculate the matrix product $A = B \cdot D^{-1},$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.lmul_D! — Function

lmul_D!(d, M)

In-place calculation of the matrix product $M = D \cdot M,$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.rmul_D! — Function

rmul_D!(M, d)

In-place calculation of the matrix product $M = M \cdot D,$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.ldiv_D! — Function

ldiv_D!(d, M)

In-place calculation of the matrix product $M = D^{-1} \cdot M,$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.rdiv_D! — Function

rdiv_D!(M, d)

In-place calculation of the matrix product $M = M \cdot D^{-1},$ where $D$ is a diagonal matrix represented by the vector d.

StableLinearAlgebra.mul_P! — Function

mul_P!(A, p, B)

Evaluate the matrix product $A = P \cdot B,$ where $P$ is a permutation matrix represented by the vector of integers p.

mul_P!(A, B, p)

Evaluate the matrix product $A = B \cdot P,$ where $P$ is a permutation matrix represented by the vector of integers p.

StableLinearAlgebra.inv_P! — Function

inv_P!(p⁻¹, p)

Calculate the inverse/transpose $P^{-1}=P^T$ of a permuation $P$ represented by the vector p, writing the result to p⁻¹.

StableLinearAlgebra.perm_sign — Function

perm_sign(p::AbstractVector{Int})

Calculate the sign/parity of the permutation p, $\textrm{sgn}(p) = \pm 1.$

LAPACK LinearAlgebra

StableLinearAlgebra.QRWorkspace — Type

QRWorkspace{T<:Number, E<:Real}

Allocated space for calcuating the pivoted QR factorization using the LAPACK routines geqp3! and orgqr! while avoiding dynamic memory allocations.

StableLinearAlgebra.LUWorkspace — Type

LUWorkspace{T<:Number, E<:Real}

Allocated space for calcuating the pivoted QR factorization using the LAPACK routine getrf!. Also interfaces with the getri! and getrs! routines for inverting matrices and solving linear systems respectively.

StableLinearAlgebra.inv_lu! — Function

inv_lu!(A::AbstractMatrix{T}, ws::LUWorkspace{T}) where {T}

Calculate the inverse of the matrix A, overwriting A in-place. Also return $\log(|\det A^{-1}|)$ and $\textrm{sign}(\det A^{-1}).$

StableLinearAlgebra.ldiv_lu! — Function

ldiv_lu!(A::AbstractMatrix{T}, B::AbstractMatrix{T}, ws::LUWorkspace{T}) where {T}

Calculate $B:= A^{-1} B,$ modifying $B$ in-place. The matrix $A$ is over-written as well. Also return $\log(|\det A^{-1}|)$ and $\textrm{sign}(\det A^{-1}).$

StableLinearAlgebra.det_lu! — Function

det_lu!(A::AbstractMatrix{T}, ws::LUWorkspace{T}) where {T}

Return $\log(|\det A|)$ and $\textrm{sign}(\det A).$ Note that $A$ is left modified by this function.